Fixed preview collisions
Building previews now have their collisions disabled, and then reenabled once the building is placed. This is in response to an issue relating to the player being able to propel themselves by continuously moving the preview to their current location. The player tag has also been added to the blacklisted tags for building placement.